database-service/requirements.txt
# Third-party imports
flask
neo4j

# OpenTelemetry imports
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
